+++ server/src/controllers/categorias.controller.js
@@ -1,4 +1,5 @@
 const { prisma } = require("../lib/prisma");
+const { HttpError } = require("../utils/http-error");
 const { TIPOS_CATEGORIA, requireBoolean, requireEnum, requireIdParam, requireString } = require("../utils/validators");
 
 async function listCategorias() {
@@ -42,9 +43,27 @@ async function updateCategoria(id, payload) {
 }
 
 async function deleteCategoria(id) {
+  const categoriaId = requireIdParam(id);
+  const [lancamentosCount, categoriasSubcontasCount] = await Promise.all([
+    prisma.lancamento.count({
+      where: {
+        categoria_id: categoriaId
+      }
+    }),
+    prisma.categoriaSubconta.count({
+      where: {
+        categoria_id: categoriaId
+      }
+    })
+  ]);
+
+  if (lancamentosCount > 0 || categoriasSubcontasCount > 0) {
+    throw new HttpError(409, "Nao e possivel excluir esta categoria porque existem lancamentos ou subcontas vinculados a ela.");
+  }
+
   await prisma.categoria.delete({
     where: {
-      id: requireIdParam(id)
+      id: categoriaId
     }
   });
